refactor : move utility related classes into util package
