+++ src/main/java/coz/weavon/validation/validator/RangeValidator.java
@@ -1,7 +1,7 @@
 package coz.weavon.validation.validator;
 
 import coz.weavon.exception.model.ClientException;
-import coz.weavon.common.application.util.DateTimeUtils;
+import coz.weavon.utils.DateTimeUtils;
 import coz.weavon.validation.model.ValidRange;
 import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
